MD5 Checksum, output log, script tidy up
Tidied the script a bit including some formatting. Added MD5 Checksum
functionality. In a previous version FCIV.exe was used to calculate MD5
Checksum but in this version, Powershell was used. Created an output log
to time how long it took and provide statistics on any fields that
needed to be modified.